[maintenance]: moving auth module
