+++ src/modules/auth/auth-controller/auth.controller.ts
@@ -2,7 +2,7 @@ import * as jwt from 'jsonwebtoken';
 import { Response } from 'express';
 import { Controller, Body, Res, Post } from '@nestjs/common';
 import { AuthService } from '../auth-service/auth.service';
-import { UserCridentials } from '../../models/user-cridentials.model';
+import { UserCridentials } from '../../../models/user-credentials.model';
 
 @Controller()
 export class AuthController {
